# Garbage collect synchronized items from channels put/get queues.