Fixes #12.