@cwebber 2017-08-14
Before:

```
cwebber@oolong:~/devel/fibers$ time ./env guile benchmarks/local-ping-pong.scm 100 10000

real	0m9.825s
user	0m18.126s
sys	0m0.240s
cwebber@oolong:~/devel/fibers$ time ./env guile benchmarks/local-ping-pong.scm 100 10000

real	0m10.057s
user	0m17.905s
sys	0m0.224s
cwebber@oolong:~/devel/fibers$ time ./env guile benchmarks/local-ping-pong.scm 100 10000

real	0m9.658s
user	0m17.825s
sys	0m0.271s
```

After:

```
cwebber@oolong:~/devel/fibers$ time ./env guile benchmarks/local-ping-pong.scm 100 10000

real	0m10.263s
user	0m18.904s
sys	0m0.251s
cwebber@oolong:~/devel/fibers$ time ./env guile benchmarks/local-ping-pong.scm 100 10000

real	0m10.526s
user	0m19.233s
sys	0m0.318s
cwebber@oolong:~/devel/fibers$ time ./env guile benchmarks/local-ping-pong.scm 100 10000

real	0m10.234s
user	0m18.870s
sys	0m0.288s
```

Looks to be about a 5% decrease in performance, CPU-wise.  To me this seems worth it to curb the nonstop memory growth.

BTW I agree that the atomic box overhead seems bothersome (as I said in my last patch too)... eventually I'd like to explore other solutions, but I think for now it's most important to get this fix in, since we do want to let people run long-lived processes without exhausting their memory!